refactor: dataloader
