+++ examples/doremi/doremi/dataloader.py
@@ -7,7 +7,7 @@ import numpy as np
 import torch
 from nanotron import distributed as dist
 from nanotron import logging
-from nanotron.dataloader import get_dataloader_worker_init
+from nanotron.data.dataloader import get_dataloader_worker_init
 from nanotron.parallel import ParallelContext
 from nanotron.parallel.pipeline_parallel.tensor_pointer import TensorPointer
 from nanotron.parallel.pipeline_parallel.utils import get_input_output_pp_ranks
